consolidate hyperbridge subxt config in a crate
