+++ Cargo.toml
@@ -111,6 +111,7 @@ sp-block-builder = { version = "26.0.0", default-features = false }
 sp-consensus-aura = { version = "0.32.0", default-features = false }
 sp-consensus-beefy = { version = "13.0.0", default-features = false }
 sp-core = { version = "28.0.0", default-features = false }
+sp-core-hashing = { version = "15.0.0", default-features = false }
 sp-inherents = { version = "26.0.0", default-features = false }
 sp-offchain = { version = "26.0.0", default-features = false }
 sp-runtime = { version = "31.0.1", default-features = false }
